refactor: Reactor is own package
